% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{bincdf}
\alias{bincdf}
\title{Binomial Cumulative Distribution Function}
\usage{
bincdf(n = 1, p = 0.5)
}
\arguments{
\item{n}{a non-negative integer.}

\item{p}{A value between 0 and 1, inclusive.}
}
\value{
y(i) = P(X <= i-1), i=1,2,...,n+1, where X is a  Binomial(n,p) rv.
}
\description{
Binomial Cumulative Distribution Function
}
\examples{
bincdf(n = 10, p = 0.5)
}
