% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{binpmf}
\alias{binpmf}
\title{Binomial Probability Mass Function}
\usage{
binpmf(n = 1, p = 0.5)
}
\arguments{
\item{n}{a non-negative integer.}

\item{p}{A value between 0 and 1, inclusive.}
}
\value{
y(i) = P(X = i-1),  i=1,...n+1, where X  is a  Binomial(n,p) rv.
}
\description{
Binomial Probability Mass Function
}
\examples{
bincdf(n = 10, p = 0.5)
}
