% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmcfpt}
\alias{dtmcfpt}
\title{First Passage Times}
\usage{
dtmcfpt(Tr, P)
}
\arguments{
\item{P}{a stochastic matrix of size N by N}

\item{T}{row vector representing the set of target states}
}
\value{
y=[y0 y1 y2], where y0, y1, and y2 are column vectors.  [y1(i) y2(i)] is the [mean,  second moment] of the first passage time to visit any of the states in the target set of states T, starting in a non-target state y0(i) for a DTMC X_n, n >= 0 with a transition probability matrix P.
}
\description{
First Passage Times
}
\examples{
dtmcfpt(c(1,2),matrix(c(rep(1/3,9)),nrow = 3))
}
