% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmclrc}
\alias{dtmclrc}
\title{Long-run Cost Rate}
\usage{
dtmclrc(P, c)
}
\arguments{
\item{P}{an irreducible square stochastic matrix of size N by N}

\item{c}{vector of length N}
}
\value{
y is the long-run expected cost per unit time for a DTMC X_n, n >= 0 with a transition probability matrix P, and that incurs an expected cost c(i) every time it visits state i.
}
\description{
Long-run Cost Rate
}
\examples{
dtmclrc(matrix(c(0.5,0.5,0.5,0.5),nrow = 2),c(1,2))
}
