% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmcod}
\alias{dtmcod}
\title{Limiting Distribution and Occupancy Distribution.}
\usage{
dtmcod(P)
}
\arguments{
\item{P}{an aperiodic irreducible square stochastic matrix of size N by N.}
}
\value{
y(i) = limit as n goes to infinity of P(X_n = i),  1 <= i <= N, where X_n, n >= 0 is an irreducible aperiodic DTMC with transition probability matrix P. If the DTMC is periodic y is the occupancy distribution of X
}
\description{
Limiting Distribution and Occupancy Distribution.
}
\examples{
dtmcod(matrix(c(0.5,0.5,0.5,0.5),nrow = 2))
}
