% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmcot}
\alias{dtmcot}
\title{Occupancy Times}
\usage{
dtmcot(P, n)
}
\arguments{
\item{P}{a square stochastic matrix of size N by N.}

\item{n}{a nonnegative integer}
}
\value{
M(i,j) = the expected number of visits to state j starting from state i by a DTMC X_n, n >= 0 with transition probability matrix P.
}
\description{
Occupancy Times
}
\examples{
dtmcot(matrix(c(0.5,0.5,0.5,0.5),nrow = 2), 5)
}
