% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmctc}
\alias{dtmctc}
\title{Total Cost}
\usage{
dtmctc(P, c, n)
}
\arguments{
\item{P}{a square stochastic matrix of size N by N}

\item{c}{vector of length N}

\item{n}{a nonnegative integer}
}
\value{
y(i) is the total expected cost incurred over time 0 through n staring in state i, for a DTMC X_n, n >= 0 with a transition probability matrix P, and that incurs an expected cost c(i) every time it visits state i.
}
\description{
Total Cost
}
\examples{
dtmctc(matrix(c(0.5,0.5,0.5,0.5),nrow = 2),c(1,2),5)
}
