% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmctd}
\alias{dtmctd}
\title{Transient Distribution}
\usage{
dtmctd(a, P, n)
}
\arguments{
\item{a}{a row vector of length N}

\item{P}{a square stochastic matrix of size N by N.}

\item{n}{a non-negative integer}
}
\value{
y(i) = P(X_n = i),  1 <= i <= N, where X_n, n >= 0 is a DTMC with transition probability matrix P and initial distribution a.
}
\description{
Transient Distribution
}
\examples{
dtmctd(c(1,1),matrix(c(0.5,0.5,0.5,0.5),nrow = 2),5)
}
