% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{dtmctdplot}
\alias{dtmctdplot}
\title{Time Plot of Transient Distribution}
\usage{
dtmctdplot(a, j, NN, p)
}
\arguments{
\item{a}{a row vector of length N}

\item{j}{a integer 1 <= j <= N}

\item{NN}{a non-negative integer}

\item{P}{a square stochastic matrix of size N by N.}
}
\value{
y(i) = P(X_n = i),  1 <= i <= N, where X_n, n >= 0 is a DTMC with transition probability matrix P and initial distribution a.
}
\description{
Time Plot of Transient Distribution
}
\examples{
dtmctdplot(c(1,1),2,5,matrix(c(0.5,0.5,0.5,0.5),nrow = 2))
}
