% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{erlangcdf}
\alias{erlangcdf}
\title{Erlang Cumulative Distribution Function}
\usage{
erlangcdf(k = 1, l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{k}{ingeter >= 1}

\item{l}{integer >= 0}

\item{x}{a row vector}
}
\value{
y(i) =  F(x(i)), where F is an Erlang(k,l) cdf.
}
\description{
Erlang Cumulative Distribution Function
}
\examples{
erlangcdf(1, 1, c(0.4,0.3,0.2,0.1))
}
