% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{erlangcdfplot}
\alias{erlangcdfplot}
\title{Plot of Erlang Cumulative Distribution Function}
\usage{
erlangcdfplot(k = 1, l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{k}{ingeter >= 1}

\item{l}{integer >= 0}

\item{x}{a row vector}
}
\value{
Plot of the cdf of Eralng(k,l) (y vs. x).
}
\description{
Plot of Erlang Cumulative Distribution Function
}
\examples{
erlangcdfplot(1, 1, c(0.4,0.3,0.2,0.1))
}
