% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{erlangpdf}
\alias{erlangpdf}
\title{Eralng Probability Density Function}
\usage{
erlangpdf(k = 1, l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{k}{ingeter >= 1}

\item{l}{>= 0}

\item{x}{a non negative row vector}
}
\value{
y(i) =f(x(i)), where f is an Erlang(k,l) pdf.
}
\description{
Eralng Probability Density Function
}
\examples{
erlangpdf(1, 1, c(0.4,0.3,0.2,0.1))
}
