% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{erlangpdfplot}
\alias{erlangpdfplot}
\title{Plot of Erlang Probability Density Function}
\usage{
erlangpdfplot(k = 1, l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{k}{ingeter >= 1}

\item{l}{>= 0}

\item{x}{a row vector}
}
\value{
Plot of the pdf of Eralng(k,l) (y vs. x).
}
\description{
Plot of Erlang Probability Density Function
}
\examples{
erlangpdfplot(1, 1, c(0.4,0.3,0.2,0.1))
}
