% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5inv}
\alias{ex5inv}
\title{Inventory Systems}
\usage{
ex5inv(s, S, y)
}
\arguments{
\item{s}{base stock level (stock not allowed to go below this level)}

\item{S}{Restocking level}

\item{y}{row vector of the pmf of weekly demand}
}
\value{
P = transition probability matrix for the inventory system problem. (See Example 5.6)
}
\description{
Inventory Systems
}
\examples{
ex5inv(5,10,c(0.4,0.3,0.2,0.1))
}
