% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5invcost}
\alias{ex5invcost}
\title{Inventory Systems Cost Model}
\usage{
ex5invcost(s, S, y, hc, ps, oc)
}
\arguments{
\item{s}{base stock level (stock not allowed to go below this level);}

\item{S}{Restocking level;}

\item{y}{row vector of the pmf of weekly demand;}

\item{hc}{cost of holding one item for one unit of time;}

\item{ps}{profit from selling one item;}

\item{oc}{cost of placing an order;}
}
\value{
c(i) = expected cost in the current period if the inventory at the beginning is i.
}
\description{
Inventory Systems Cost Model
}
\examples{
ex5invcost(5,10,c(0.4,0.3,0.2,0.1),1,5,2)
}
