% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5manp}
\alias{ex5manp}
\title{Manpower Systems}
\usage{
ex5manp(p, l, a)
}
\arguments{
\item{p}{a row vector, p(i) =  probability of promotion from grade i to i+1}

\item{l}{a row vector of the same length as p,  l(i) =  probability of leaving from grade i}

\item{a}{row vector, a(i) = probability that a new employee joins grade i.}
}
\value{
P = transition  probability matrix for the manpower system problem. (See Example 5.8)
}
\description{
Manpower Systems
}
\examples{
ex5manp(c(0.4,0.4,0.2,0),c(0.4,0.3,0.2,0.1),c(0.4,0.3,0.2,0.1))
}
