% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5manpcost}
\alias{ex5manpcost}
\title{Manpower Systems Cost Model}
\usage{
ex5manpcost(p, l, a, s, b, d, tr)
}
\arguments{
\item{p}{a row vector, p(i) =  probability of promotion from grade i to i+1. The last element of p must be zero.}

\item{l}{a row vector of the same length as p, l(i) =  probability of leaving from grade i.}

\item{a}{row vector, a(i) = probability that a new employee joins grade i. Must be a valid pmf.}

\item{s}{a row vector, s(i) = salary of an employee in grade i.}

\item{b}{a row vector, b(i) = bonus for promotion from grade i to i+1. The last element of b must be zero.}

\item{d}{a row vector, d(i) = cost of an employee departing from grade i.}

\item{t}{a row vector, t(i) = cost of training an employee staring in grade i.}
}
\value{
c = a row vector, c(i) = expected one-period cost in state i.
}
\description{
Manpower Systems Cost Model
}
\examples{
ex5manpcost(c(0.4,0.4,0.2,0),c(0.4,0.3,0.2,0.1),c(0.4,0.3,0.2,0.1),c(1,2,3,4),c(1,2,3,0),c(1,2,3,4),c(1,2,3,4))
}
