% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5mfgcost}
\alias{ex5mfgcost}
\title{Manufacturing Systems Cost Model}
\usage{
ex5mfgcost(A, B, a1, a2, r, du, hA, hB)
}
\arguments{
\item{A}{size of bin for machine 1;}

\item{B}{size of bin for machine 2;}

\item{a1}{prob(non-defective) for machine 1;}

\item{a2}{prob(non-defective) for machine 2.}

\item{r}{revenue from a complete assembly;}

\item{du}{cost of turning a machine on;}

\item{hA}{cost of holding an item in bin A for one period;}

\item{hB}{cost of holding an item in bin B for one period;}
}
\value{
c = row vector, c(i) = expected one-period cost in state i.
}
\description{
Manufacturing Systems Cost Model
}
\examples{
ex5mfgcost(5,3,0.5,0.3,10,1,2,1)
}
