% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5mr}
\alias{ex5mr}
\title{Machine Reliability}
\usage{
ex5mr(uu, dd, k, r)
}
\arguments{
\item{uu}{P(up|up)}

\item{dd}{P(down|down)}

\item{k}{number of machines}

\item{r}{number of repair persons}
}
\value{
P = transition  probability matrix for the machine reliability problem. (See Example 5.4)
}
\description{
Machine Reliability
}
\examples{
ex5mr(0.5,0.5,3,1)
}
