% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5mrcost}
\alias{ex5mrcost}
\title{Machine Reliability Cost Model}
\usage{
ex5mrcost(uu, dd, k, r, ru, cd, cbr)
}
\arguments{
\item{uu}{P(machine is up tomorrow|it is up today)}

\item{dd}{P(machine is down tomorrow|it is down today)}

\item{k}{number of machines}

\item{r}{number of repair persons}

\item{ru}{per day revenue of an up machine}

\item{cd}{per day cost of a down machine}

\item{cbr}{per day cost of a busy repair person.}
}
\value{
c(i) = one day cost if i machines are working at the beginning of the day.
}
\description{
Machine Reliability Cost Model
}
\examples{
ex5mrcost(0.5,0.5,3,1,10,5,3)
}
