% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5stock}
\alias{ex5stock}
\title{Stock Market}
\usage{
ex5stock(L, U)
}
\arguments{
\item{L}{Lower bound for the stock price,}

\item{U}{Upper bound for the stock price.}
}
\value{
P = transition  probability matrix for the stock market.
}
\description{
Stock Market
}
\examples{
ex5stock(1,10)
}
