% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5tel}
\alias{ex5tel}
\title{Telecommunications}
\usage{
ex5tel(K, a)
}
\arguments{
\item{K}{buffer capacity}

\item{a}{= row vector, a(i) =  p(i-1 packets arrive during one time slot)}
}
\value{
P = transition  probability matrix for the Telecommunications system. (See  Example 5.10.)
}
\description{
Telecommunications
}
\examples{
ex5tel(5,c(0.1,0.2,0.3,0.3,0.1))
}
