% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMC.R
\name{ex5telcost}
\alias{ex5telcost}
\title{Telecommunications Cost Model}
\usage{
ex5telcost(K, a, rt, cl)
}
\arguments{
\item{K}{= buffer capacity,}

\item{a}{= row vector, a(i) =  p(i-1 packets arrive during one time slot). a must be a valid pmf.}

\item{rt}{=  revenue from transmitting a single packet,}

\item{cl}{=  cost of losing a single packet.}
}
\value{
c = column vector, c(i) =  expected cost in one slot if there are i-1 packets in the buffer at the end of the previous slot, 1 <= i <= K+1.
}
\description{
Telecommunications Cost Model
}
\examples{
ex5telcost(5,c(0.1,0.2,0.3,0.3,0.1),10,1)
}
