% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{expcdf}
\alias{expcdf}
\title{Exponential Cumulative Distribution Function}
\usage{
expcdf(l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{l}{>= 0}

\item{x}{a non negative row vector}
}
\value{
y(i) =  F(x(i)), where F is an Exp(l) cdf
}
\description{
Exponential Cumulative Distribution Function
}
\examples{
expcdf(1,c(0.4,0.3,0.2,0.1))
}
