% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{exppdf}
\alias{exppdf}
\title{Title Exponential Probability Density Function}
\usage{
exppdf(l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{l}{>= 0}

\item{x}{a non negative row vector}
}
\value{
y(i) =f(x(i)), where f is an Exp(l) pdf.
}
\description{
Title Exponential Probability Density Function
}
\examples{
exppdf(1,c(0.4,0.3,0.2,0.1))
}
