% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{exppdfplot}
\alias{exppdfplot}
\title{Plot of Exponential Probability Density Function}
\usage{
exppdfplot(l = 1, x = c(0.4, 0.3, 0.2, 0.1))
}
\arguments{
\item{l}{>= 0}

\item{x}{a non negative row vector}
}
\value{
Plot of the pdf of exp(l) (y vs. x).
}
\description{
Plot of Exponential Probability Density Function
}
\examples{
exppdfplot(1,c(0.4,0.3,0.2,0.1))

}
