% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{geometriccdf}
\alias{geometriccdf}
\title{Geometric Cumulative Distribution Function}
\usage{
geometriccdf(p = 0.5, k = 5)
}
\arguments{
\item{p}{value between 0 and 1, inclusive}

\item{k}{integer >= 1}
}
\value{
y(i) = P(X <= i),  i=1,2,...,k, where X  is a Geometric(p) rv.
}
\description{
Geometric Cumulative Distribution Function
}
\examples{
geometriccdf(0.5,5)
}
