% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{geometricpmf}
\alias{geometricpmf}
\title{Geometric Probability Mass Function}
\usage{
geometricpmf(p = 0.5, k = 5)
}
\arguments{
\item{p}{value between 0 and 1, inclusive}

\item{k}{integer >= 1}
}
\value{
y(i) = P(X = i),  i=1,...k, where X is a  Geometric(p) rv.
}
\description{
Geometric Probability Mass Function
}
\examples{
geometricpmf(0.5,5)
}
