% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{negbinpmf}
\alias{negbinpmf}
\title{Negative Binomial Probability Mass Function}
\usage{
negbinpmf(r = 1, p = 0.5, k = 1)
}
\arguments{
\item{r}{integer >= 1}

\item{p}{value between 0 and 1, inclusive}

\item{k}{integer >= 1}
}
\value{
y(i) = P(X = i),  i=r,r+1,...,r+k, where X  is a  Negative Binomial(r,p) rv.
}
\description{
Negative Binomial Probability Mass Function
}
\examples{
negbinpmf(1,0.5,1)
}
