% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{normalcdf}
\alias{normalcdf}
\title{Normal Cumulative Distribution Function}
\usage{
normalcdf(m = 1, s = 1, x = c(1, 2, 3, 4))
}
\arguments{
\item{m}{mean of the distribution}

\item{s}{variance, which should be bigger than 0}

\item{x}{a row vector}
}
\value{
y(i) =  F(x(i)), where F is an Normal(m,s) cdf, with mean  m and variance s.
}
\description{
Normal Cumulative Distribution Function
}
\examples{
normalcdf(1,1,c(1,2,3,4))
}
