% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{normalcdfplot}
\alias{normalcdfplot}
\title{Plot of Normal Cumulative Distribution Function}
\usage{
normalcdfplot(m = 1, s = 1, x = c(1, 2, 3, 4))
}
\arguments{
\item{m}{mean of the distribution}

\item{s}{variance, which should be bigger than 0}

\item{x}{a row vector}
}
\value{
Plot of the cdf of Normal(m,s) (y vs. x).
}
\description{
Plot of Normal Cumulative Distribution Function
}
\examples{
normalcdfplot(1,1,c(1,2,3,4))
}
