% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{normalpmf}
\alias{normalpmf}
\title{Normal Probability Density Function}
\usage{
normalpmf(m = 1, s = 1, x = c(1, 2, 3, 4))
}
\arguments{
\item{m}{mean of the distribution}

\item{s}{variance, which should be bigger than 0}

\item{x}{a row vector}
}
\value{
y(i) =f(x(i)), where f is an Normal(m,s) pdf, with mean m and variance s.
}
\description{
Normal Probability Density Function
}
\examples{
normalpdf(1,1,c(1,2,3,4))
}
