% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{normalpmfplot}
\alias{normalpmfplot}
\title{Plot of Normal Probability Density Function}
\usage{
normalpmfplot(m = 1, s = 1, x = c(1, 2, 3, 4))
}
\arguments{
\item{m}{mean of the distribution}

\item{s}{variance, which should be bigger than 0}

\item{x}{a row vector}
}
\value{
Plot of the pdf of Normal(m,s) (y vs. x).
}
\description{
Plot of Normal Probability Density Function
}
\examples{
normalpdfplot(1,1,c(1,2,3,4))
}
