% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{poissoncdf}
\alias{poissoncdf}
\title{Poisson Cumulative Distribution Function}
\usage{
poissoncdf(l, k)
}
\arguments{
\item{l}{value of the poisson parameter}

\item{k}{maximum value of X}
}
\value{
y(i) = P(X <= i-1),  i=1,2,...,k+1, where X is a  Poisson(l) rv.
}
\description{
Poisson Cumulative Distribution Function
}
\examples{
poissoncdf(1,5)
}
