% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{poissonpmf}
\alias{poissonpmf}
\title{Poisson Probability Mass Function}
\usage{
poissonpmf(l, k)
}
\arguments{
\item{l}{value of the poisson parameter}

\item{k}{maximum value of X}
}
\value{
y(i) = P(X = i-1),  i=1,...k+1, where X  is a Poisson(l) rv.
}
\description{
Poisson Probability Mass Function
}
\examples{
poissonpmf(1,5)
}
